ディーラーとやり取りする部分以外をコンポーネントに寄せる
